Align use of '-' and '--' with other CLI parsers
The use a single '-' is usually for short options a single char. Also
the way it is used in all tests. One of the example has an alias "ag"
which currently is parsed if the argument is '-ag' with this change
aliases with more the a single char need to be parsed with double
dash - as e.g. '--ag'