Align flag alias with normal use (#85)
* Align use of '-' and '--' with other CLI parsers

The use a single '-' is usually for short options a single char. Also
the way it is used in all tests. One of the example has an alias "ag"
which currently is parsed if the argument is '-ag' with this change
aliases with more the a single char need to be parsed with double
dash - as e.g. '--ag'

* Normalize args '-abcd=10' as '-a -b -c -d 10'

This is also the normal behavior of CLI arg parsers - you can group
mutilple short options with a single dash.

* Refactor 'normalized_args' to common 'utils.rs'

* Bump msrv rust 1.35.0 -> 1.67.0

* Add test

* Handle both short and long aliases